==> update/notebooks/1_get_articles.ipynb <==
import os, io, json, urllib, numpy as np, codecs, os.path,time
this_month='1710'
data_path='../data/'+this_month

def get_data(cursor,i,year):
    global item, all_items, url
    item={}; items=[]; 
    p={ 'filter':'from-pub-date:'+year+',until-pub-date:'+year , 'rows' : '1000', 'cursor' :cursor}
    parameters=urllib.urlencode(p)
    url='http://api.crossref.org/types/journal-article/works?'+parameters
    response = urllib.urlopen(url).read()
    json_result=json.loads(response)
    items = json_result['message']['items']
    if len(items)>0:
        f=codecs.open(data_path+'/'+year+'/'+str(i)+'.txt','a','utf-8')
        for item in items:
            if 'DOI' in item and 'issn-type' in item:
                journal=pub=cite=ref=issnp=issne=lic=''; year_print=9999;year_online=9999                                
                for i in item['issn-type']:
                    if i['type']=='print': 
                        issnp=i['value']
                    elif i['type']=='electronic':
                        issne=i['value']
                doi=item['DOI']
                year=str(item['issued']['date-parts'][0])
                cite=str(item['is-referenced-by-count'])
                if 'publisher' in item: pub=item['publisher']; pub='' if pub==None else pub
                if 'container-title' in item: journal=','.join(item['container-title'])
                if 'license' in item: 
                    lic=item['license'][0]['URL'].split('://')[1].split('/')[0]
                res= doi+'\t'+year+'\t'+cite+'\t'+issnp+'\t'+issne+'\t'+journal+'\t'+pub+'\t'+lic+'\n'
                f.write(res)    
        f.close()
    next_cursor=json_result['message']['next-cursor']
    return next_cursor

 for year in range(2014,2018):
    year=str(year)
    if not os.path.exists('../data/'+year):
        os.makedirs(data_path+'/'+year)    
    cursor=''
    next_cursor='*'
    item={}
    all_items=[]
    i=0
    while next_cursor!=cursor:
        cursor=next_cursor
        while True:
            try:
                next_cursor=get_data(cursor,i,year)
            except Exception,e: 
                print str(e)
                time.sleep(1)
                continue
            break
        i=i+1
        if i%10==0:print i,

url

cursor
