==> update/notebooks/2_get_journals.ipynb <==
import os, io, json, urllib, numpy as np, codecs
import pandas as pd
import pickle


this_month='1710'
data_path='../data/'+this_month

base_url="https://api.crossref.org/journals?rows=1000&offset="

url="https://api.crossref.org/journals?rows=1"
response = urllib.urlopen(url).read()
json_result=json.loads(response)
count=json_result['message']['total-results']
dic=[]
for it in range(count/1000 + 1):
    print it,
    offset = it * 1000
    url=base_url+str(offset)
    response = urllib.urlopen(url).read()
    json_result=json.loads(response)
    items=json_result['message']['items']
    for item in items:
        if len(item['ISSN'])>0 and 'title' in item and 'publisher' in item: 
            journal=dict([
                    ('title',item['title'].strip()),
                    ('publisher',item['publisher']),
                    ('issn',item['ISSN']),
                    ('subjects',[subject['name'].strip() for subject in item['subjects']])
                ])
            dic.append(journal)
dic_c=dic    

with open(data_path +'/journals.pkl', 'wb') as f:
    pickle.dump(dic, f, pickle.HIGHEST_PROTOCOL)
