==> update/notebooks/3_merge_articles.ipynb <==
%matplotlib inline
import os, io, json, urllib, numpy as np, codecs
import glob, pandas as pd
pd.options.display.max_rows = 999

this_month='1710'
data_path='../data/'+this_month
columns=8

y17 = glob.glob(data_path+'/2017/*.txt')
y16 = glob.glob(data_path+'/2016/*.txt')
y15 = glob.glob(data_path+'/2015/*.txt')
y14 = glob.glob(data_path+'/2014/*.txt')
y14_17=y14+y15+y16+y17
with open(data_path+'/merged.txt', "wb") as outfile:
    for f in y14_17:
        with open(f, "rb") as infile:
            lines=infile.read().split('\n')
            for line in lines:
                if len(line.split('\t'))==columns:
                    outfile.write(line+'\n')

df=pd.read_table(data_path+'/merged.txt',header=None)
df.columns=['doi','year','citation','issnp','issne','journal','publisher','lic']
df.shape

df=df[(~df.issnp.isnull()) | (~df.issne.isnull())]
df.shape

df['count']=1
df['cited']=df['citation'].apply(lambda x: 1 if x>0 else 0)
df['un-cited']=df['citation'].apply(lambda x: 0 if x>0 else 1)

df['mm']=df.year.apply(lambda x: int(x.replace(']','').split(', ')[1]) if len(x.split(', '))>1 else 0)

df['year']=df.year.str[1:5].astype(int)

df=df.fillna('_')
df['issn']=df.issnp+','+df.issne
df['issn']=df.issn.str.replace(',_','')
df['issn']=df.issn.str.replace('_,','')
df['issn']=df.issn.apply(lambda x:x[0:9] if x[10:19]==x[0:9] else x)

df.to_csv(data_path+'/cleaned.csv',index=False)
