==> update/notebooks/4_journal_stats.ipynb <==
%matplotlib inline
import os, io, json, urllib, numpy as np, codecs
import glob, pandas as pd
pd.options.display.max_rows = 999
import pickle

this_month='1710'
month_number=int(this_month[2:])
data_path='../data/'+this_month

def h_(citation):
    citation=list(citation)
    h=0
    citation.sort(reverse=True)
    for i in range(1,len(citation)+1):
        if citation[i-1] >= i:
            h=i
    return h

df=pd.read_csv(data_path+'/cleaned.csv')
df.shape

df=df[~((df.year==2014) & (df.mm<month_number))]
df.shape

df=df[(df['year']<=2017) & (df['year']>=2014)]
df.shape

def yymm(yy,mm):
    yymm=(yy%100)*100+mm
    if yymm<1500+month_number: return 2014
    if yymm<1600+month_number: return 2015
    if yymm<1700+month_number: return 2016

df['yymm']=df.apply(lambda x: yymm(x['year'],x['mm']),1)

with open(data_path +'/journals.pkl', 'rb') as f:
    journals_dic=pickle.load(f)

issn_journal={}
for journal in journals_dic:
    for issn in journal['issn']:
        issn_journal[issn]=journal

df['issns']=df.apply(lambda x: issn_journal[x['issnp']]['issn'] if x['issnp'] in issn_journal else (issn_journal[x['issne']]['issn'] if x['issne'] in issn_journal else '-'),1)

df['@id']=df.issns.apply(lambda x:'_'.join(x))
df.shape

df=df[df['@id']!='-']
df.shape

dg=df.groupby(['@id','yymm']).agg({'count':sum,'citation':sum}).reset_index()
dg['impact']=dg['citation'].div(dg['count']).round(2)
dg.shape

dg_h=df[['@id','citation']].groupby(['@id']).agg(lambda x:tuple(x)).reset_index()
dg_h['h']=dg_h.citation.apply(lambda x: h_(x))
dg_h.shape

dg_i=df.groupby(['@id']).agg({'count':sum,'citation':sum}).reset_index()
dg_i['impact']=dg_i['citation'].div(dg_i['count']).round(2)
dg_i.shape

dg_flat=dg.pivot(index='@id',columns='yymm')
dg_flat.columns = [col[0]+'_'+str(col[1])[3] for col in dg_flat.columns.values]
dg_flat=dg_flat.reset_index()

dg_merge=dg_flat.merge(dg_i[['@id','count','citation','impact']])\
.merge(dg_h[['@id','h']]).fillna(0,downcast='infer')
dg_merge.shape

unique_subjects=[]
for journal in journals_dic:
    for s in journal['subjects']: 
        if s not in unique_subjects: 
            unique_subjects.append(s)
len(unique_subjects)  

pd.DataFrame(unique_subjects).to_csv('../store/unique_subjects.csv',header=False,index=False,encoding='utf-8')

with open('..\web\cats.json') as data_file:    
    cat_subject=json.load(data_file)

sub_cat={}
for cat in cat_subject.keys():
    for sub in cat_subject[cat].split(';'):
        sub_cat[sub]=cat

dic_new=[]
for j in journals_dic:
    j_new=j
    j_new['category']=[]
    for subject in j['subjects']:
        cat=sub_cat[subject]
        if cat not in j_new['category']:
            j_new['category'].append(sub_cat[subject])
    dic_new.append(j_new)  
len(dic_new)        

with open('../store/road.tsv') as f:
    road_lines = f.read().splitlines()
road_lines=road_lines[1:]

road_urls=[issn[1:-1] for issn in road_lines]
road_issns=[url[-9:] for url in road_urls]

def issns_open(issns):
    for issn in issns.split('_'):
        if issn in road_issns:
            return road_urls[road_issns.index(issn)]

dg_merge['road_url']=dg_merge['@id'].apply(lambda x:issns_open(x))

issns_journal={}
for journal in dic_new:
    issns_journal['_'.join(journal['issn'])]=journal

dg_merge['title']=dg_merge['@id'].apply(lambda x:issns_journal[x]['title'])
dg_merge['publisher']=dg_merge['@id'].apply(lambda x:issns_journal[x]['publisher'])
dg_merge['category']=dg_merge['@id'].apply(lambda x:issns_journal[x]['category'])
dg_merge['subjects']=dg_merge['@id'].apply(lambda x:issns_journal[x]['subjects'])

work_batches=dg_merge.shape[0]/1000
for s in range(work_batches+1):
    dic=[]
    for index, i in dg_merge[1000*s:1000*(s+1)].iterrows():
        journal={}
        journal['@id']=i['@id']
        for y in range(4,7): 
            journal['e:count_'+str(y)]=i['count_'+str(y)]
            journal['e:citation_'+str(y)]=i['citation_'+str(y)]
            journal['e:impact_'+str(y)]=round(i['impact_'+str(y)],2)
        journal['e:count']=i['count']
        journal['e:citation']=i['citation']
        journal['e:impact']=round(i['impact'],2)
        journal['e:title']=i['title']
        journal['e:publisher']=i['publisher']
        journal['e:h']=i['h']        
        if len(i['category'])>0: journal['e:category']=i['category'] 
        if len(i['subjects'])>0: journal['e:subjects']=i['subjects'] 
        if i['road_url']!=None: journal['e:open']=1
        if i['road_url']!=None: journal['e:road_url']=i['road_url']
        if journal not in dic:
            dic.append(journal)
    with open('../jsonld/'+this_month+'/'+'journals_'+str(s)+'.jsonld', 'w') as outfile:
        json.dump(dic, outfile)

dg_merge[['@id','subjects']].to_csv('../store/issn_subject.csv',index=False)
