==> update/notebooks/5_category_subjects.ipynb <==
import os, io, json, urllib, numpy as np, codecs
import pandas as pd
from ast import literal_eval

unique_subjects=pd.read_csv('../store/unique_subjects.csv',header=None)
unique_subjects=unique_subjects[0].values

with open ('../store/ASJC.txt','r') as myfile:
    ASJC=myfile.read()

ASJC=ASJC.replace('General','Multidisciplinary')
ASJC=ASJC.replace('veterinary','Veterinary')
ASJC=ASJC.replace('General veterinary','General Veterinary')
ASJC=ASJC.replace('veterinary (miscalleneous','Veterinary (miscalleneous)')

from StringIO import StringIO
subject_cats=pd.read_table(StringIO(ASJC),delimiter=';')#from github
subject_cats.shape 
subject_cats['Low']=subject_cats.Low.apply(lambda x: 'General '+x.replace('(all)','') if '(all)' in x else x)

for i in subject_cats.Low.values:
    if i not in unique_subjects:
        print i

for i in unique_subjects:
    if i not in subject_cats.Low.values:
        print i

cats={}
for i, row in subject_cats.iterrows():
    cat=row.Middle.strip()
    subject=(row.Low).strip()
    if cat in cats:
        cats[cat]=cats[cat]+';'+subject
    else:
        cats[cat]=subject

with open('..\web\cats.json', 'w') as outfile:
    json.dump(cats, outfile)
