==> update/notebooks/6_top20.ipynb <==
%matplotlib inline
import os, io, json, urllib, numpy as np, codecs
import glob, pandas as pd
from ast import literal_eval
pd.options.display.max_rows = 999
import pickle

this_month='1710'
month_number=int(this_month[2:])
data_path='../data/'+this_month

df=pd.read_csv(data_path+'/cleaned.csv')
df.shape

df=df[~((df.year==2014) & (df.mm<month_number))]
df.shape

df=df[(df['year']<=2017) & (df['year']>=2014)]
df.shape

with open(data_path +'/journals.pkl', 'rb') as f:
    journals_dic=pickle.load(f)

df['issns']=df.apply(lambda x: issn_journal[x['issnp']]['issn'] if x['issnp'] in issn_journal else (issn_journal[x['issne']]['issn'] if x['issne'] in issn_journal else '-'),1)

df2=df[['doi','issns','citation']]

df2['@id']=df2.issns.apply(lambda x:'_'.join(x))

issn_subjects=pd.read_csv('../store/issn_subject.csv')
issn_subjects.columns=['@id','subjects']
issn_subjects.subjects=issn_subjects['subjects'].apply(literal_eval)

df2=df2.merge(issn_subjects)

unique_subjects=[]
all_subjects=df2['subjects'].values

unique_subjects=set(x for l in all_subjects for x in l)

len(unique_subjects)

top_articles=[]
counter=0
for subject in unique_subjects:
    if subject=='':
        df_subject=df2.sort_values('citation', ascending= False)[:20]
    else:
        df_subject=df2[df2.subjects.apply(lambda x: subject in x)].sort_values('citation', ascending= False)[:20]
    top_dois=df_subject[['doi','@id','citation']].to_dict('list')
    zips=zip(top_dois['doi'],top_dois['@id'],top_dois['citation'])
    items=[];
    for doi,issn,citation in zips:
        item={}; item_cleaned={}
        url='http://api.crossref.org/works/'+doi
        while True:
            try:
                response = urllib.urlopen(url).read()    
            except Exception,e: 
                print str(e)
                time.sleep(1)
                continue
            break
        json_result=json.loads(response)
        item = json_result['message']
        year_print=[9999,99,99];year_online=[9999,99,99];
        if 'title' in item:
            title=item['title'][0]
        if 'published-print' in item:
            year_print=item['published-print']['date-parts'][0]                
        if 'published-online' in item:
            year_online=item['published-online']['date-parts'][0]
        year=min(year_print,year_online)
        if 'DOI' in item: doi=item['DOI']
        if 'container-title' in item: journal=','.join(item['container-title'])
        if 'publisher' in item: pub=item['publisher']; pub=''if pub==None else pub
        if 'references-count' in item: ref=str(item['references-count'])
        authors=[]
        if 'author' in item: 
            for author in item['author']:
                if 'given' in author:
                    author_name=author['given']
                if 'family' in author:
                    author_name+=' '+author['family']
                authors.append(author_name)
        if len(authors)>5: authors=authors[:5]
        if len(year)>1 :
            month=str(year[1]) 
        else: month=''
        item_cleaned={'@id':doi,
                      'e:title':title,
                      'e:year':str(year[0])+'-'+month,
                      'e:cite':np.asscalar(citation),
                      'e:issn':issn, 
                      'e:journal':journal,
                      'e:pub':pub,
                      'e:authors':authors,
                      'e:subject':subject,
                      'e:month':this_month} 
        if item_cleaned not in top_articles:
            top_articles.append(item_cleaned)
    print counter,
    counter+=1            

with open('..\web\cats.json') as data_file:    
    cat_subject=json.load(data_file)

sub_cat={}
for cat in cat_subject.keys():
    for sub in cat_subject[cat].split(';'):
        sub_cat[sub]=cat

top_articles_new=[]
for a in top_articles:
    a['e:subject']=a['e:subject'].strip()
    a['e:category']=''
    subject=a['e:subject']
    if subject!='':
        cat=sub_cat[subject]
        a['e:category']=sub_cat[subject]
    top_articles_new.append(a)

with open('../jsonld/'+this_month+'/articles.jsonld', 'w') as fp:
    json.dump(top_articles_new, fp)
